@@ -18,18 +18,19 @@
   },
   "homepage": "https://github.com/tylermcginnis/github-notetaker-egghead",
   "dependencies": {
-    "axios": "^0.5.4",
-    "firebase": "^2.2.7",
-    "re-base": "^0.1.1",
-    "react": "^0.13.3",
-    "react-router": "^0.13.3",
-    "reactfire": "^0.4.0"
+    "axios": "^0.13.1",
+    "firebase": "^3.3.0",
+    "re-base": "^2.0.0",
+    "react": "^15.3.0",
+    "react-dom": "^15.3.0",
+    "react-router": "^2.6.1",
+    "reactfire": "^1.0.0"
   },
   "devDependencies": {
-    "babel-core": "^5.6.5",
-    "babel-loader": "^5.1.4",
-    "http-server": "^0.8.5",
-    "node-libs-browser": "^0.5.2",
-    "webpack": "^1.9.11"
+    "babel-core": "^5.8.38",
+    "babel-loader": "^5.4.2",
+    "http-server": "^0.9.0",
+    "node-libs-browser": "^1.0.0",
+    "webpack": "^1.13.1"
   }
 }
