@@ -1,7 +1,5 @@
 import React from 'react';
-import Router from 'react-router';
+import { render } from 'react-dom';
 import routes from './config/routes';
 
-Router.run(routes, (Root, state) => {
-  React.render(<Root {...state} />, document.getElementById('app'));
-});
+render(routes, document.getElementById('app'))
